% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{bill_calendar}
\alias{bill_calendar}
\title{Bill Calendar}
\usage{
bill_calendar(bill)
}
\arguments{
\item{bill}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{scalar-type-predicates}}, \code{\link[rlang]{is_integerish}}, \code{\link[rlang]{abort}}
\code{\link[purrr]{pluck}}
}
