% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/billids.R
\name{billids}
\alias{billids}
\title{Page validation}
\usage{
billids(billobject)
}
\arguments{
\item{billobject}{The page value passed from the legiscan function}
}
\value{
Returns a valid page object for inclusion in the URL specification
}
\description{
Checks for valid values of page parameter for search method
}
\examples{
\dontrun{
# Pass an unprocessed/cleaned bill object to billids to clean up the
# identifiers in the API response
billobject_cleaner <- billids(billobject_raw)
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{fileLists}};
  \code{\link{fileStructure}}; \code{\link{legiBillHist}};
  \code{\link{legiBillJSON}}; \code{\link{legiBillMeta}};
  \code{\link{legiBillProg}}; \code{\link{legiBillSpons}};
  \code{\link{legiBillTxt}}; \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

