% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caller-generics.R, R/caller-methods.R
\docType{methods}
\name{billText}
\alias{billText}
\alias{billText,LegiScan,numeric-method}
\title{LegiScan Caller Methods - billText}
\usage{
billText(legiscan, id)

\S4method{billText}{LegiScan,numeric}(legiscan, id)
}
\arguments{
\item{legiscan}{An object of class LegiScan}

\item{id}{A LegiScan doc_id value}
}
\description{
Generic method for the LegiScan getBillText API call

Retrieves the full bill text, metadata (e.g., draft revision number, etc...),
and MIME type for the text.  The text is encoded in base64.
}
\seealso{
Other LegiScan API Caller Generics: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,NULL,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,NULL,NULL-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,numeric-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,NULL,NULL,NULL-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,NULL,NULL-method}},
  \code{\link{masterList,LegiScan,character,NULL-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,NULL,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,NULL,NULL-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,numeric-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,NULL,NULL,NULL-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,NULL,NULL-method}},
  \code{\link{masterList,LegiScan,character,NULL-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}
}

