% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkPage.R
\name{checkPage}
\alias{checkPage}
\title{Page validation}
\usage{
checkPage(x)
}
\arguments{
\item{x}{The page value passed from the legiscan function}
}
\value{
Returns a valid page object for inclusion in the URL specification
}
\description{
Checks for valid values of page parameter for search method
}
\examples{
\dontrun{
page <- checkPage(page)
}
}

