% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkState.R
\name{checkState}
\alias{checkState}
\title{State validation}
\usage{
checkState(x)
}
\arguments{
\item{x}{The state value passed from the legiscan function}
}
\value{
Returns a valid state object for inclusion in the URL specification
}
\description{
Checks for valid values of state parameter for legiscan methods
}
\examples{
\dontrun{
state <- checkState(state)
}
}

