% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkYear.R
\name{checkYear}
\alias{checkYear}
\title{Year validation}
\usage{
checkYear(x)
}
\arguments{
\item{x}{The year value passed from the legiscan function}
}
\value{
Returns a valid year object for inclusion in the URL specification
}
\description{
Checks for valid values of year parameter for search method
}
\examples{
\dontrun{
year <- checkYear(year)
}
}

