% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cleanBill}
\alias{cleanBill}
\title{Wrappers - Parsing/Cleaning Bill Data}
\usage{
cleanBill(BillObject, data = FALSE, text = "")
}
\arguments{
\item{BillObject}{An object returned from the bill method for the LegiScan class}

\item{data}{A logical indicating whether the function should return
separate data frame objects (TRUE) or return a list of data frames (FALSE)}

\item{text}{A character parameter passed to the parseBill method to identify
if the function should attempt to retrieve the full text of the bill}
}
\value{
Returns the texts object with the full text of the bill(s) joined to
the data frame object
}
\description{
This is a wrapper function for the parseBill method
}
\seealso{
Other LegiScan Wrapper Functions: \code{\link{cleanAmendment}};
  \code{\link{cleanBillText}};
  \code{\link{cleanMasterList}};
  \code{\link{cleanRollCall}}; \code{\link{cleanSessions}};
  \code{\link{cleanSponsor}}; \code{\link{cleanStates}};
  \code{\link{cleanSupplement}}
}

