% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cleanBillText}
\alias{cleanBillText}
\title{Wrappers - Parsing/Cleaning API response to the getBillText call}
\usage{
cleanBillText(BillText)
}
\arguments{
\item{rawBillText}{Object returned from the billText method for the LegiScan class}
}
\value{
Returns a list containing meta data and the binary MIME object containing
the full text of the bill
}
\description{
This is a wrapper function for the parseBillText method
}
\note{
It is recommended to retrieve the full text of the bill by specifying a
value (e.g., 'url' or 'state_link') to the cleanBill function if you are only
interested in the text itself.
}
\seealso{
Other LegiScan Wrapper Functions: \code{\link{cleanAmendment}};
  \code{\link{cleanBill}}; \code{\link{cleanMasterList}};
  \code{\link{cleanRollCall}}; \code{\link{cleanSessions}};
  \code{\link{cleanSponsor}}; \code{\link{cleanStates}};
  \code{\link{cleanSupplement}}
}

