% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cleanMasterList}
\alias{cleanMasterList}
\title{Wrappers - Parsing/Cleaning Master Lists}
\usage{
cleanMasterList(rawMasterList, data = FALSE, makeArchive = FALSE)
}
\arguments{
\item{rawMasterList}{Object returned from the masterList method for the LegiScan class}

\item{data}{A logical indicating whether the function should return
a data frame object (TRUE) or a tbl_df object (FALSE)}

\item{makeArchive}{A logical indicating whether or not to save an archive file}
}
\value{
Returns the master list of all legislation post processing/cleaning
}
\description{
This is a wrapper function for the parseMasterList method
}
\seealso{
Other LegiScan Wrapper Functions: \code{\link{cleanAmendment}};
  \code{\link{cleanBillText}}; \code{\link{cleanBill}};
  \code{\link{cleanRollCall}}; \code{\link{cleanSessions}};
  \code{\link{cleanSponsor}}; \code{\link{cleanStates}};
  \code{\link{cleanSupplement}}
}

