% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cleanRollCall}
\alias{cleanRollCall}
\title{Wrappers - Parsing/Cleaning API response to the getRollCall call}
\usage{
cleanRollCall(aRollCall, data = FALSE)
}
\arguments{
\item{aRollCall}{Object returned from the rollCall method for the LegiScan class}

\item{data}{A logical indicating whether the data should be returned to the global
environment as individual objects (TRUE); or as a list of data frame objects (FALSE)}
}
\value{
Returns voting meta data and individual legislator voting records
}
\description{
This is a wrapper function for the parseRollCall method
}
\seealso{
Other LegiScan Wrapper Functions: \code{\link{cleanAmendment}};
  \code{\link{cleanBillText}}; \code{\link{cleanBill}};
  \code{\link{cleanMasterList}};
  \code{\link{cleanSessions}}; \code{\link{cleanSponsor}};
  \code{\link{cleanStates}}; \code{\link{cleanSupplement}}
}

