% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cleanSessions}
\alias{cleanSessions}
\title{Wrappers - Parsing/Cleaning Sessions List}
\usage{
cleanSessions(rawSessionList, data = FALSE)
}
\arguments{
\item{rawSessionList}{Object returned from the sessions method for the LegiScan class}

\item{data}{A logical indicating whether the function should return
a data frame object (TRUE) or return a list (FALSE)}
}
\value{
Returns the texts object with the full text of the bill(s) joined to
the data frame object
}
\description{
This is a wrapper function for the parseSessions method
}
\seealso{
Other LegiScan Wrapper Functions: \code{\link{cleanAmendment}};
  \code{\link{cleanBillText}}; \code{\link{cleanBill}};
  \code{\link{cleanMasterList}};
  \code{\link{cleanRollCall}}; \code{\link{cleanSponsor}};
  \code{\link{cleanStates}}; \code{\link{cleanSupplement}}
}

