% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cleanSponsor}
\alias{cleanSponsor}
\title{Wrappers - Parsing/Cleaning API response to the getSponsor call}
\usage{
cleanSponsor(aSponsor, data = FALSE)
}
\arguments{
\item{aSponsor}{Object returned from the sponsor method for the LegiScan class}

\item{data}{A logical indicating whether to return a data.frame object (true)
or a tbl_df object (FALSE)}
}
\value{
Returns ID data for an individual bill sponsor
}
\description{
This is a wrapper function for the parseSponsor method
}
\seealso{
Other LegiScan Wrapper Functions: \code{\link{cleanAmendment}};
  \code{\link{cleanBillText}}; \code{\link{cleanBill}};
  \code{\link{cleanMasterList}};
  \code{\link{cleanRollCall}}; \code{\link{cleanSessions}};
  \code{\link{cleanStates}}; \code{\link{cleanSupplement}}
}

