% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cleanStates}
\alias{cleanStates}
\title{Wrappers - Parsing/Cleaning State List}
\usage{
cleanStates(StateList, data = FALSE)
}
\arguments{
\item{StateList}{Object returned from the stateList method for the LegiScan class}

\item{data}{A logical indicating whether the function should return
a data frame object (TRUE) or return a list (FALSE)}
}
\value{
Returns the texts object with the full text of the bill(s) joined to
the data frame object
}
\description{
This is a wrapper function for the parseStates method
}
\seealso{
Other LegiScan Wrapper Functions: \code{\link{cleanAmendment}};
  \code{\link{cleanBillText}}; \code{\link{cleanBill}};
  \code{\link{cleanMasterList}};
  \code{\link{cleanRollCall}}; \code{\link{cleanSessions}};
  \code{\link{cleanSponsor}}; \code{\link{cleanSupplement}}
}

