% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cleanSupplement}
\alias{cleanSupplement}
\title{Wrappers - Parsing/Cleaning API response to the getSupplement call}
\usage{
cleanSupplement(aSupplement)
}
\arguments{
\item{anSupplement}{Object returned from the supplement method for the LegiScan class}
}
\value{
Returns a list containing meta data and the binary MIME object containing
the full text of the supplement
}
\description{
This is a wrapper function for the parseSupplement method
}
\seealso{
Other LegiScan Wrapper Functions: \code{\link{cleanAmendment}};
  \code{\link{cleanBillText}}; \code{\link{cleanBill}};
  \code{\link{cleanMasterList}};
  \code{\link{cleanRollCall}}; \code{\link{cleanSessions}};
  \code{\link{cleanSponsor}}; \code{\link{cleanStates}}
}

