% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fileLists.R
\name{fileLists}
\alias{fileLists}
\title{Build file listing for Legiscan data}
\usage{
fileLists(fileobject)
}
\arguments{
\item{fileobject}{The object produced by the returned value from fileStructure()}
}
\value{
Creates a list object containing three nested lists for each
legislative session
}
\description{
Creates an object used to pass filenames to other legiscan functions
}
\details{
\itemize{
	\item{"bills"}{list of full filepaths/names for all bills XML/JSON files}
	\item{"people"}{list of full filepaths/names for all people XML/JSON files}
	\item{"votes"}{list of full filepaths/names for all vote  XML/JSON files}
	}
}
\examples{
\donttest{
# Create an object to store the directory tree
directoryTree <- fileStructure("data/msHistorical/")

# Build a file object
files <- fileLists(directoryTree)
}
}
\seealso{
Other Legiscan Data Downloads: \code{\link{fileStructure}}

Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileStructure}}; \code{\link{legiBillHist}};
  \code{\link{legiBillJSON}}; \code{\link{legiBillMeta}};
  \code{\link{legiBillProg}}; \code{\link{legiBillSpons}};
  \code{\link{legiBillTxt}}; \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

