% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fileStructures.R
\name{fileStructure}
\alias{fileStructure}
\title{Build Directory Tree Object for Legiscan Data Downloads}
\usage{
fileStructure(filepath)
}
\arguments{
\item{filepath}{The file path to the root directory of the legiscan data}
}
\value{
Creates a list object containing three nested lists for each
directory in the path
}
\description{
Creates an object to store all of the directory/file information
required to build the fileLists object and parse the legiscan data
}
\details{
\itemize{
	\item{"bills"}{list of full filepaths/names for all bills XML/JSON files}
	\item{"people"}{list of full filepaths/names for all people XML/JSON files}
	\item{"votes"}{list of full filepaths/names for all vote  XML/JSON files}
	}
}
\examples{
\donttest{
# Build a directory tree object
directoryTree <- fileStructure("data/msHistorical/")
}
}
\seealso{
Other Legiscan Data Downloads: \code{\link{fileLists}}

Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{legiBillHist}};
  \code{\link{legiBillJSON}}; \code{\link{legiBillMeta}};
  \code{\link{legiBillProg}}; \code{\link{legiBillSpons}};
  \code{\link{legiBillTxt}}; \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

