% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{get_master_list_raw}
\alias{get_master_list_raw}
\title{Get Master List Raw}
\usage{
get_master_list_raw(id = NULL, state = NULL)
}
\arguments{
\item{id}{A session ID}

\item{state}{A state abbreviation}
}
\description{
Get a master list of all bills, optionally for a specific state.
}
