% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{get_roll_call}
\alias{get_roll_call}
\title{Get Roll Call}
\usage{
get_roll_call(id)
}
\arguments{
\item{id}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
