% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{get_session_list}
\alias{get_session_list}
\title{Get Session List}
\usage{
get_session_list(state = NULL)
}
\arguments{
\item{state}{A state abbreviation}
}
\description{
Get a list of all sessions, optionally for a specific state.
}
