% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessor-generics.R, R/accessor-methods.R
\docType{methods}
\name{getAPI}
\alias{getAPI}
\alias{getAPI,LegiScan-method}
\alias{getAPI,LegiScan-methods}
\title{LegiScan Accessor Methods - getAPI}
\usage{
getAPI(legiscan)

\S4method{getAPI}{LegiScan}(legiscan)
}
\arguments{
\item{legiscan}{An object of class LegiScan}
}
\value{
Returns the users' api key stored in the LegiScan object
}
\description{
Generic accessor method for retrieving the value from the api slot

Accessor method for the api slot of the LegiScan class object
}
\examples{
\dontrun{
# Create an object of class LegiScan
myLegiScan <- legiscanR()

# Retrieve the value of the stored API key
getAPI(myLegiScan)
}
}
\seealso{
Other LegiScan Accessor Generics: \code{\link{getFormat}},
  \code{\link{getFormat,LegiScan-method}},
  \code{\link{getFormat,LegiScan-methods}};
  \code{\link{getUrl}},
  \code{\link{getUrl,LegiScan-method}},
  \code{\link{getUrl,LegiScan-methods}}

Other LegiScan Accessor Methods: \code{\link{getFormat}},
  \code{\link{getFormat,LegiScan-method}},
  \code{\link{getFormat,LegiScan-methods}};
  \code{\link{getUrl}},
  \code{\link{getUrl,LegiScan-method}},
  \code{\link{getUrl,LegiScan-methods}}
}

