% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessor-generics.R, R/accessor-methods.R
\docType{methods}
\name{getFormat}
\alias{getFormat}
\alias{getFormat,LegiScan-method}
\alias{getFormat,LegiScan-methods}
\title{LegiScan Accessor Methods - getFormat}
\usage{
getFormat(legiscan)

\S4method{getFormat}{LegiScan}(legiscan)
}
\arguments{
\item{legiscan}{An object of class LegiScan}
}
\value{
Returns the value stored in format slot of the LegiScan object
}
\description{
Generic accessor method retrieving the format slot of a
LegiScan class object

Accessor method for the format slot of the LegiScan class object
}
\examples{
\dontrun{
# Create object of class LegiScan
myLegiScan <- legiscanR()

# Retrieve the
getFormat(myLegiScan)
}
}
\seealso{
Other LegiScan Accessor Generics: \code{\link{getAPI}},
  \code{\link{getAPI,LegiScan-method}},
  \code{\link{getAPI,LegiScan-methods}};
  \code{\link{getUrl}},
  \code{\link{getUrl,LegiScan-method}},
  \code{\link{getUrl,LegiScan-methods}}

Other LegiScan Accessor Methods: \code{\link{getAPI}},
  \code{\link{getAPI,LegiScan-method}},
  \code{\link{getAPI,LegiScan-methods}};
  \code{\link{getUrl}},
  \code{\link{getUrl,LegiScan-method}},
  \code{\link{getUrl,LegiScan-methods}}
}

