% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFullText.R
\name{getFullText}
\alias{getFullText}
\title{Full Bill Text Retreival}
\usage{
getFullText(data, link = c("state_link", "url"))
}
\arguments{
\item{data}{The texts data object produced internally from the parseBill method}

\item{link}{Argument defines where the function will attempt to retrieve the
text from}
}
\value{
Returns the texts object with the full text of the bill(s) joined to
the data frame object
}
\description{
Used to retrieve the full text of legislation from LegiScan and/or the official
State URL provided by the LegiScan API call
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{legiBillHist}}; \code{\link{legiBillJSON}};
  \code{\link{legiBillMeta}}; \code{\link{legiBillProg}};
  \code{\link{legiBillSpons}}; \code{\link{legiBillTxt}};
  \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

