% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessor-generics.R, R/accessor-methods.R
\docType{methods}
\name{getUrl}
\alias{getUrl}
\alias{getUrl,LegiScan-method}
\alias{getUrl,LegiScan-methods}
\title{LegiScan Accessor Methods - getUrl}
\usage{
getUrl(legiscan)

\S4method{getUrl}{LegiScan}(legiscan)
}
\arguments{
\item{legiscan}{An object of class LegiScan}
}
\value{
Returns the value stored in urlBase slot of the LegiScan object
}
\description{
Generic accessor method for retrieving the base of the request URL

Accessor method for retrieving the urlBase slot of
a LegiScan class object
}
\examples{
\dontrun{
# Create object of class LegiScan
myLegiScan <- legiscanR()

# Retrieve the base URL from the LegiScan object
getUrl(myLegiScan)
}
}
\seealso{
Other LegiScan Accessor Generics: \code{\link{getAPI}},
  \code{\link{getAPI,LegiScan-method}},
  \code{\link{getAPI,LegiScan-methods}};
  \code{\link{getFormat}},
  \code{\link{getFormat,LegiScan-method}},
  \code{\link{getFormat,LegiScan-methods}}

Other LegiScan Accessor Methods: \code{\link{getAPI}},
  \code{\link{getAPI,LegiScan-method}},
  \code{\link{getAPI,LegiScan-methods}};
  \code{\link{getFormat}},
  \code{\link{getFormat,LegiScan-method}},
  \code{\link{getFormat,LegiScan-methods}}
}

