% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiBillHist.R
\name{legiBillHist}
\alias{legiBillHist}
\title{Legiscan Bill Historical Data}
\usage{
legiBillHist(fileobject)
}
\arguments{
\item{fileobject}{A JSON file object from the bills subdirectory}
}
\value{
Creates a list object containing a data frame
for the history of the bill
}
\description{
Parses and arranges JSON output from
Legiscan master data downloads bills subdirectory
}
\examples{
\donttest{
# Create the directory tree in an object
directoryTree <- fileStructure("data/msHistoricalJSON/")

# Store the file path/structure in a new object to reference files for parsing
files <- fileLists(directoryTree)

# Parse the bill history data from a bill file
history <- legiBillHist(files[["bills"]][[10]][[12]])
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{legiBillJSON}}; \code{\link{legiBillMeta}};
  \code{\link{legiBillProg}}; \code{\link{legiBillSpons}};
  \code{\link{legiBillTxt}}; \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

