% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiBillJSON.R
\name{legiBillJSON}
\alias{legiBillJSON}
\title{Legiscan Bill Data}
\usage{
legiBillJSON(file)
}
\arguments{
\item{file}{A JSON file object from the vote subdirectory}
}
\value{
Creates a list object containing a data frame
for voting metadata and individual legislators' votes
}
\description{
Parses and arranges JSON output from
Legiscan master data dumps vote subdirectory
}
\examples{
\donttest{
# Create directory tree object
directoryTree <- fileStructure("data/msHistorical/")

# Build file list object
files <- fileLists(directoryTree)

# Parse/clean up the Bill data
bills <- legiBillJSON(files[["bills"]][[1]][[1]])
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{legiBillHist}}; \code{\link{legiBillMeta}};
  \code{\link{legiBillProg}}; \code{\link{legiBillSpons}};
  \code{\link{legiBillTxt}}; \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

