% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiBillMeta.R
\name{legiBillMeta}
\alias{legiBillMeta}
\title{Legiscan Bill Metadata}
\usage{
legiBillMeta(file)
}
\arguments{
\item{file}{A JSON file object from the bills subdirectory}
}
\value{
Creates a list object containing a data frame
of metadata for an individual piece of legislation
}
\description{
Parses and arranges JSON output from
Legiscan master data downloads bills subdirectory
}
\examples{
\donttest{
# Create directory object
directoryTree <- fileStructure("data/msHistorical/")

# Create file list object
files <- fileLists(directoryTree)

# Parse the meta data from the bill file
bills <- legiBillMeta(files[["bills"]][[1]][[1]])
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{getFullText}}; \code{\link{legiBillHist}};
  \code{\link{legiBillJSON}}; \code{\link{legiBillProg}};
  \code{\link{legiBillSpons}}; \code{\link{legiBillTxt}};
  \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

