% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiBillProg.R
\name{legiBillProg}
\alias{legiBillProg}
\title{Legiscan Bill Progress Tracking Data}
\usage{
legiBillProg(fileobject)
}
\arguments{
\item{fileobject}{A JSON file object from the bills subdirectory}
}
\value{
Creates a list object containing a data frame
for the progress of the bill
}
\description{
Parses and arranges JSON output from
Legiscan master data downloads bills subdirectory
}
\examples{
\donttest{
# Create directory object
directoryTree <- fileStructure("data/msHistoricalJSON/")

# Create file list object
files <- fileLists(directoryTree)

# Parse/clean the bill progress data from a LegiScan bill file
billProgress <- legiBillProg(files[["bills"]][[10]][[12]])
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{legiBillHist}}; \code{\link{legiBillJSON}};
  \code{\link{legiBillMeta}}; \code{\link{legiBillSpons}};
  \code{\link{legiBillTxt}}; \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

