% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiBillSpons.R
\name{legiBillSpons}
\alias{legiBillSpons}
\title{Legiscan Bill Sponsor Data}
\usage{
legiBillSpons(fileobject)
}
\arguments{
\item{fileobject}{A JSON file object from the bills subdirectory}
}
\value{
Creates a list object containing a data frame
for the progress of the bill
}
\description{
Parses and arranges JSON output from
Legiscan master data downloads bills subdirectory
}
\examples{
\donttest{
# Create directory tree object
directoryTree <- fileStructure("data/msHistorical/")

# Build file list object
files <- fileLists(directoryTree)

# Parse/clean up the Bill Sponsor data
billSponsors <- legiBillSpons(files[["bills"]][[10]][[12]])
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{legiBillHist}}; \code{\link{legiBillJSON}};
  \code{\link{legiBillMeta}}; \code{\link{legiBillProg}};
  \code{\link{legiBillTxt}}; \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

