% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiPersonJSON.R
\name{legiPersonJSON}
\alias{legiPersonJSON}
\title{Legiscan Legislator Data}
\usage{
legiPersonJSON(file)
}
\arguments{
\item{file}{A JSON file object from the people subdirectory}
}
\value{
Creates a list object containing a data frame
for identifying individual legislators
}
\description{
Parses and arranges JSON data from
Legiscan master data downloads people subdirectory
}
\examples{
\donttest{
# Create directory object
directoryTree <- fileStructure("data/msHistoricalJSON/")

# Create file list object
files <- fileLists(directoryTree)

# Parse/clean the full bill text data from a LegiScan bill file
aLegislator <- legiPersonJSON(files[["people"]][[10]][[12]])
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{legiBillHist}}; \code{\link{legiBillJSON}};
  \code{\link{legiBillMeta}}; \code{\link{legiBillProg}};
  \code{\link{legiBillSpons}}; \code{\link{legiBillTxt}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

