% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiscan-class.R
\docType{class}
\name{LegiScan-class}
\alias{LegiScan}
\alias{LegiScan-class}
\alias{initialize,LegiScan-method}
\alias{legiscanR}
\title{LegiScan API Class object}
\usage{
legiscanR(...)

\S4method{initialize}{LegiScan}(.Object, urlBase, api, format)
}
\arguments{
\item{...}{Should be passed arguments named by the slot names from getSlots("LegiScan")}

\item{.Object}{a LegiScan object to create}

\item{urlBase}{Root URL for LegiScan API calls}

\item{api}{If specified, will search for matching file otherwise will treat as character string}

\item{format}{XML or JSON API requests (set on the LegiScan website)}
}
\value{
Returns the LegiScan object if the validation tests are passed
}
\description{
Defines a LegiScan class object which is passed to methods
to submit/retrieve data from calls to the LegiScan API and for parsing
data retrieve from the LegiScan API

LegiScan constructor function

LegiScan default construction method
}
\section{Slots}{

\describe{
\item{\code{urlBase}}{The base for the request URL (predefined in the class prototype)}

\item{\code{api}}{The API key for the end user}

\item{\code{format}}{Specifies if you have requested your API pulls in XML or JSON}
}}
\note{
The default behavior for the class constructor is to read the API key
from ~/.legiscanAPI.  Accessor methods are available to get/set the values
of the LegiScan object.

The file format returned by the LegiScan API is set up when logging
into your LegiScan account.  The value provided by the class object is used
to select the appropriate parsing methods based on the object returned
from the API call
}
\examples{
\donttest{
# Create new object with class LegiScan
x <- legiscanR()

# Check class of the object x
class(x)
}
}

