% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiscanBill.R
\name{legiscanBill}
\alias{legiscanBill}
\title{Legiscan Bill Data}
\usage{
legiscanBill(file)
}
\arguments{
\item{file}{An XML file object from the bill subdirectory of the master download}
}
\value{
Creates a list object containing a several data.frame objects.
The objects include the bill's meta data, status, history, text, and
vote histories.
}
\description{
Parses and arranges XML output from
Legiscan master data dumps vote subdirectory
}
\examples{
\donttest{
# Build directory structure from master data downloads
directoryTree <- fileStructure("data/msHistorical/")

# Create object with all file references stored in it
files <- fileLists(directoryTree)

# Pass the function a bill object for processing
bills <- legiscanBill(files[["bills"]][[1]][[1]])
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{getFullText}}; \code{\link{legiBillHist}};
  \code{\link{legiBillJSON}}; \code{\link{legiBillMeta}};
  \code{\link{legiBillProg}}; \code{\link{legiBillSpons}};
  \code{\link{legiBillTxt}}; \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

