% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiscanR.R
\docType{package}
\name{legiscanR-package}
\alias{legiscanR-package}
\title{legiscanR functions and the LegiScan Class object/methods}
\description{
An R interface to the LegiScan API (\url{http://legiscan.com/}) for
accessing US legislative data.  This package provides the LegiScan class
and methods for sending requests to the API for regularly updated
changes to legislation/votes, as well as functions and methods to
parse the data from the API calls and from master download files.
}

