% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caller-generics.R, R/caller-methods.R
\docType{methods}
\name{legisearch}
\alias{legisearch}
\alias{legisearch,LegiScan,character,missing,character,missing,missing-method}
\alias{legisearch,LegiScan,character,missing,character,missing,numeric-method}
\alias{legisearch,LegiScan,character,missing,character,numeric,missing-method}
\alias{legisearch,LegiScan,character,missing,character,numeric,numeric-method}
\alias{legisearch,LegiScan,character,numeric,missing,missing,missing-method}
\title{LegiScan Caller Methods - legisearch}
\usage{
legisearch(legiscan, state, bill, query, year, page)


  \S4method{legisearch}{LegiScan,character,numeric,missing,missing,missing}(legiscan,
  state, bill, query, year, page)


  \S4method{legisearch}{LegiScan,character,missing,character,missing,missing}(legiscan,
  state, bill, query, year, page)


  \S4method{legisearch}{LegiScan,character,missing,character,numeric,missing}(legiscan,
  state, bill, query, year, page)


  \S4method{legisearch}{LegiScan,character,missing,character,missing,numeric}(legiscan,
  state, bill, query, year, page)


  \S4method{legisearch}{LegiScan,character,missing,character,numeric,numeric}(legiscan,
  state, bill, query, year, page)
}
\arguments{
\item{legiscan}{An object of class LegiScan}

\item{state}{A two character US State abbreviation (validated against the
state.abb dataset in R)}

\item{bill}{A LegiScan billnumber for an exact bill number or to limit query}

\item{query}{A character vector containing the search terms}

\item{year}{An indicator for either classes of years or for a specific year
from 1900 through till the present}

\item{page}{Indicates the page number of the results requested}
}
\description{
Generic method for the LegiScan search API call

Generic method for the LegiScan search API call

Generic method for the LegiScan search API call

Generic method for the LegiScan search API call

Generic method for the LegiScan search API call

Generic method for the LegiScan search API call
}
\details{
The accepted values the the year parameter are described in more detail here.
\itemize{
		\item{"year = 1"}{All Years of Data Available}
		\item{"year = 2"}{Current Year of Data Only (default)}
		\item{"year = 3"}{Recent Years of Data Available}
		\item{"year = 4"}{Prior Years of Data Available}
		\item{"year >= 1900"}{Exact Year Only}
}

The accepted values the the year parameter are described in more detail here.
\itemize{
		\item{"year = 1"}{All Years of Data Available}
		\item{"year = 2"}{Current Year of Data Only (default)}
		\item{"year = 3"}{Recent Years of Data Available}
		\item{"year = 4"}{Prior Years of Data Available}
		\item{"year >= 1900"}{Exact Year Only}
}

The accepted values the the year parameter are described in more detail here.
\itemize{
		\item{"year = 1"}{All Years of Data Available}
		\item{"year = 2"}{Current Year of Data Only (default)}
		\item{"year = 3"}{Recent Years of Data Available}
		\item{"year = 4"}{Prior Years of Data Available}
		\item{"year >= 1900"}{Exact Year Only}
}

The accepted values the the year parameter are described in more detail here.
\itemize{
		\item{"year = 1"}{All Years of Data Available}
		\item{"year = 2"}{Current Year of Data Only (default)}
		\item{"year = 3"}{Recent Years of Data Available}
		\item{"year = 4"}{Prior Years of Data Available}
		\item{"year >= 1900"}{Exact Year Only}
}

The accepted values the the year parameter are described in more detail here.
\itemize{
		\item{"year = 1"}{All Years of Data Available}
		\item{"year = 2"}{Current Year of Data Only (default)}
		\item{"year = 3"}{Recent Years of Data Available}
		\item{"year = 4"}{Prior Years of Data Available}
		\item{"year >= 1900"}{Exact Year Only}
}

The accepted values the the year parameter are described in more detail here.
\itemize{
		\item{"year = 1"}{All Years of Data Available}
		\item{"year = 2"}{Current Year of Data Only (default)}
		\item{"year = 3"}{Recent Years of Data Available}
		\item{"year = 4"}{Prior Years of Data Available}
		\item{"year >= 1900"}{Exact Year Only}
}
}
\seealso{
Other LegiScan API Caller Generics: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,missing,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,missing,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,missing,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,missing,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,missing,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,missing,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}
}

