% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiVotesJSON.R
\name{legiVotesJSON}
\alias{legiVotesJSON}
\title{Legiscan Voting Data}
\usage{
legiVotesJSON(file)
}
\arguments{
\item{file}{A JSON file object from the vote subdirectory}
}
\value{
Creates a list object containing a data frame
for voting metadata and individual legislators' votes
}
\description{
Parses and arranges JSON data from
Legiscan master data dumps vote subdirectory
}
\examples{
\donttest{
# Create directory object
directoryTree <- fileStructure("data/msHistoricalJSON/")

# Create file list object
files <- fileLists(directoryTree)

# Parse/clean the voting records data from a LegiScan votes file
votes <- legiscanVotes(files[["votes"]])
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{legiBillHist}}; \code{\link{legiBillJSON}};
  \code{\link{legiBillMeta}}; \code{\link{legiBillProg}};
  \code{\link{legiBillSpons}}; \code{\link{legiBillTxt}};
  \code{\link{legiPersonJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

