% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caller-generics.R, R/caller-methods.R
\docType{methods}
\name{masterList}
\alias{masterList}
\alias{masterList,LegiScan,NULL,NULL-method}
\alias{masterList,LegiScan,character,NULL-method}
\title{LegiScan Caller Methods - masterList}
\usage{
masterList(legiscan, state, id)

\S4method{masterList}{LegiScan,character,`NULL`}(legiscan, state, id)

\S4method{masterList}{LegiScan,`NULL`,`NULL`}(legiscan, state, id)
}
\arguments{
\item{legiscan}{An object of class LegiScan}

\item{state}{A two character US State abbreviation (validated against the
state.abb dataset in R)}

\item{id}{A LegiScan session_id value}
}
\description{
Generic method for the LegiScan getMasterList API call

Method to retrieve master list of bills introduced in the current legislative
session.  The response also provides a checksum for tracking changes to the
text of bills, as well as bill_id and billnumber identifiers.

Method to retrieve master list of bills introduced in the current legislative
session.  The response also provides a checksum for tracking changes to the
text of bills, as well as bill_id and billnumber identifiers.
}
\details{
The getMasterList API call has two distinct methods available.  The method
which LegiScan recommends, is to make master list requests using the state
abbreviation.  However, a second method is available which will select the
data based on the session id number assigned by LegiScan.
}
\examples{
\dontrun{
# Construct LegiScan Class Object
myLegiScan <- legiscanR(format = "XML")

# Request the masterlist for Little Rhody
masterList(myLegiScan, "RI")
}
}
\seealso{
Other LegiScan API Caller Generics: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,NULL,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,NULL,NULL-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,numeric-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,NULL,NULL,NULL-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,NULL,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,NULL,NULL-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,numeric-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,NULL,NULL,NULL-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,NULL,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,NULL,NULL-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,NULL,numeric-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,NULL-method}},
  \code{\link{legisearch,LegiScan,character,NULL,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,NULL,NULL,NULL-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}
}

