% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{new_bill}
\alias{new_bill}
\title{New Bill}
\usage{
new_bill(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
