% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parser-generics.R, R/parser-methods.R
\docType{methods}
\name{parseBill}
\alias{parseBill}
\alias{parseBill,XMLDocumentContent,logical,character-method}
\alias{parseBill,character,logical,character-method}
\title{LegiScan Parser Methods - parseBill}
\usage{
parseBill(rawBill, dataframe = FALSE, fullText = "")

\S4method{parseBill}{XMLDocumentContent,logical,character}(rawBill,
  dataframe = FALSE, fullText = "")

\S4method{parseBill}{character,logical,character}(rawBill, dataframe = FALSE,
  fullText = "")
}
\arguments{
\item{rawBill}{An object with the returned data from the masterList method}

\item{dataframe}{If TRUE method returns multiple data objects to the
global environment; if FALSE the method returns a list of data frames}

\item{fullText}{Accepts either "", "state_link", or "url" as arguments; the ""
argument is used to skip retrieving the full text of the bill; the "state_link"
argument is used to retrieve the full text of the bill from the State's URL
provided in the API call response; the "url" argument will attempt to retrieve the
full text of the bill from LegiScan.}
}
\value{
Returns multiple data frame objects if dataframe = TRUE or returns a
list of data frames if the dataframe = FALSE

Returns multiple data frame objects if dataframe = TRUE or returns a
list of data frames if the dataframe = FALSE
}
\description{
Generic method for parsing data retrieved from getBill API call

Method for parsing ID data from LegiScan API calls to getBill
}
\examples{
\dontrun{
# Create object of class LegiScan
myLegiScan <- legiscanR()

# Get the master list to get bill IDs
master <- masterList(myLegiScan, state = "MS")

# Get the bill data for the first bill in the session
theBill <- bill(myLegiScan, master[1,3])

# Parse the XML/JSON formatted data into a data frame
parseBillIDs(theBill)

}
}
\seealso{
Other LegiScan API Parser Generics: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical,numeric-method}},
  \code{\link{parseMasterList,character,logical,logical,missing-method}},
  \code{\link{parseSessions,character,logical,logical,missing-method}};
  \code{\link{parseQuery}},
  \code{\link{parseQuery,XMLDocumentContent-method}},
  \code{\link{parseQuery,character-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}};
  \code{\link{parseSponsor}},
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates}},
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical,numeric-method}},
  \code{\link{parseMasterList,character,logical,logical,missing-method}},
  \code{\link{parseSessions,character,logical,logical,missing-method}};
  \code{\link{parseQuery}},
  \code{\link{parseQuery,XMLDocumentContent-method}},
  \code{\link{parseQuery,character-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}};
  \code{\link{parseSponsor}},
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates}},
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical,numeric-method}},
  \code{\link{parseMasterList,character,logical,logical,missing-method}},
  \code{\link{parseSessions,character,logical,logical,missing-method}};
  \code{\link{parseQuery}},
  \code{\link{parseQuery,XMLDocumentContent-method}},
  \code{\link{parseQuery,character-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}};
  \code{\link{parseSponsor}},
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates}},
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}
}

