% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parser-generics.R, R/parser-methods.R
\docType{methods}
\name{parseBillText}
\alias{parseBillText}
\alias{parseBillText,XMLDocumentContent-method}
\alias{parseBillText,character-method}
\title{LegiScan Parser Methods - parseBillText}
\usage{
parseBillText(theBillText)

\S4method{parseBillText}{XMLDocumentContent}(theBillText)

\S4method{parseBillText}{character}(theBillText)
}
\arguments{
\item{theBillText}{An object with the returned data from the billText method}
}
\value{
A list object containing metadata and the MIME data retrieved by the API call
}
\description{
Generic method for parsing data retrieved from getBillText API call

Method for parsing LegiScan API calls to getBillText
}
\note{
The getBillText API call method returns a binary MIME object with any
of several document format types.  While this may be of interest, it is much
easier to retrieve the text data from the cleanBill function with an option
specified for the text argument.
}
\examples{
\dontrun{
# Create object of class LegiScan
myLegiScan <- legiscanR()

# Get the master list for Mississippi
theMasterList <- masterList(myLegiScan, "MS")

# Parse the XML/JSON formatted data into a data frame
cleanerMasterList <- parseMasterList(theMasterList)

# Get A bill (bill_id is in the 3rd Column of the master list)
aBill <- bill(myLegiScan, cleanerMasterList[1, 3])

# Parse the bill data
cleanerBill <- cleanBill(aBill, dataframe = FALSE, text = 'state_link')

# Retrieve the Bill Text from the API call
legiscanBillText <- billText(myLegiScan, cleanerBill[["texts"]][["doc_id"]])

# Store the API response in new object
aBillText <- legiscanBillText[[1]]

# Parse the bill text response from LegiScan
cleanedBillText <- parseBillText(aBillText)

}
}
\seealso{
Other LegiScan API Parser Generics: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseQuery}}; \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}}; \code{\link{parseSponsor}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}};
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}};
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}
}

