% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parser-generics.R, R/parser-methods.R
\docType{methods}
\name{parseMasterList}
\alias{parseMasterList}
\alias{parseMasterList,XMLDocumentContent,logical,logical,numeric-method}
\alias{parseMasterList,character,logical,logical,missing-method}
\alias{parseSessions,character,logical,logical,missing-method}
\title{LegiScan Parser Methods - parseMasterList}
\usage{
parseMasterList(rawMasterList, dataframe = FALSE, archive = TRUE,
  option = c(RECOVER, NOCDATA))


  \S4method{parseMasterList}{XMLDocumentContent,logical,logical,numeric}(rawMasterList,
  dataframe = FALSE, archive = TRUE, option = c(RECOVER, NOCDATA))

\S4method{parseMasterList}{character,logical,logical,missing}(rawMasterList,
  dataframe = FALSE, archive = TRUE, option = NULL)
}
\arguments{
\item{rawMasterList}{An object with the returned data from the masterList method}

\item{dataframe}{A logical defining the returned object data type (set to
false for tbl_df object)}

\item{archive}{A logical specifying if an archived file will be created. File will
be saved as MasterListArchive followed by a timestamp in the current directory.}

\item{option}{Additional arguments that can be supplied to the XML/RJSONIO parsers}
}
\description{
Generic method for parsing data retrieved from getMasterList API call

Method for parsing LegiScan API calls to getMasterList
}
\note{
Additional arguments should only be used when formatting errors cause a
fatal error that you believe can be recovered without sacraficing your data.
For example, an invalid unicode character may appear in a single location in the
document, in the middle of the text string.  In that case, you may wish to add:
options = c(NOCDATA, NOERROR, RECOVER) to the internal call to the XML parser.
By default option = c(RECOVER, NOCDATA) is specified
}
\examples{
\dontrun{
# Create object of class LegiScan
myLegiScan <- legiscanR()

# Get the master list for Mississippi
theMasterList <- masterList(myLegiScan, "MS")

# Parse the XML/JSON formatted data into a data frame
parseMasterList(theMasterList)

# Or parse into a list
parseMasterList(theMasterList, dataframe = FALSE)

# R
}
}
\seealso{
Other LegiScan API Parser Generics: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseQuery}},
  \code{\link{parseQuery,XMLDocumentContent-method}},
  \code{\link{parseQuery,character-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}};
  \code{\link{parseSponsor}},
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates}},
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseQuery}},
  \code{\link{parseQuery,XMLDocumentContent-method}},
  \code{\link{parseQuery,character-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}};
  \code{\link{parseSponsor}},
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates}},
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseQuery}},
  \code{\link{parseQuery,XMLDocumentContent-method}},
  \code{\link{parseQuery,character-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}};
  \code{\link{parseSponsor}},
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates}},
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}
}

