% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parser-generics.R, R/parser-methods.R
\docType{methods}
\name{parseRollCall}
\alias{parseRollCall}
\alias{parseRollCall,XMLDocumentContent,logical-method}
\alias{parseRollCall,character,logical-method}
\title{LegiScan Parser Methods - parseRollCall}
\usage{
parseRollCall(theRollCall, dataframe = FALSE)

\S4method{parseRollCall}{XMLDocumentContent,logical}(theRollCall,
  dataframe = FALSE)

\S4method{parseRollCall}{character,logical}(theRollCall, dataframe = FALSE)
}
\arguments{
\item{theRollCall}{An object with the returned data from the rollCall method}

\item{dataframe}{If TRUE method returns multiple data objects to the
global environment; if FALSE the method returns a list of data frames}
}
\description{
Generic method for parsing data retrieved from getRollCall API call

Method for parsing LegiScan API calls to getRollCall
}
\note{
The getRollCall API call method returns a binary MIME object with any
of several document format types.  Amendments and rollCalls are not supported
in all states.
}
\examples{
\dontrun{
# Create object of class LegiScan
myLegiScan <- legiscanR()

# Get the master list for Mississippi
theMasterList <- masterList(myLegiScan, "MS")

# Parse the XML/JSON formatted data into a data frame
cleanerMasterList <- parseMasterList(theMasterList)

# Get A bill (bill_id is in the 3rd Column of the master list)
aBill <- bill(myLegiScan, cleanerMasterList[1, 3])

# Parse the bill data
cleanerBill <- cleanBill(aBill, dataframe = FALSE, text = '')

# Retrieve the RollCall from the API call
aRollCall <- rollCall(myLegiScan, cleanerBill[["votes"]][["roll_call_id"]])

# Parse the bill text response from LegiScan
cleanedRollCall <- parseRollCall(aRollCall)

}
}
\seealso{
Other LegiScan API Parser Generics: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseQuery}}; \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}}; \code{\link{parseSponsor}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}};
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}};
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}
}

