% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parser-generics.R, R/parser-methods.R
\docType{methods}
\name{parseStates}
\alias{parseSessions}
\alias{parseSessions,XMLDocumentContent,logical-method}
\alias{parseSessions,character,logical-method}
\alias{parseStates}
\title{LegiScan Parser Methods - parseStates}
\usage{
parseStates(rawStateList, dataframe = TRUE)

parseSessions(rawSessionList, dataframe = TRUE)

\S4method{parseSessions}{XMLDocumentContent,logical}(rawSessionList,
  dataframe = TRUE)

\S4method{parseSessions}{character,logical}(rawSessionList, dataframe = TRUE)
}
\arguments{
\item{rawStateList}{An object with the returned data from the stateList method}

\item{dataframe}{A logical defining the returned object data type (set to false for a list)}

\item{rawSessionList}{An object with the returned data from the sessionList method}

\item{dataframe}{A logical defining the returned object data type (set to false for a list)}
}
\description{
Generic method for parsing data retrieved from stateList API call

Generic method for parsing data retrieved from sessionList API call

Method for parsing LegiScan API calls to getSessionList
}
\examples{
\dontrun{
# Create object of class LegiScan
myLegiScan <- legiscanR()

# Get the session list for Mississippi
theSessions <- sessionList(myLegiScan, "MS")

# Parse the XML/JSON formatted data into a data frame
parseSessions(theSessions)

# Or parse into a list
parseSessions(theSessions, dataframe = FALSE)
}
}
\seealso{
Other LegiScan API Parser Generics: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseQuery}}; \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSponsor}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan API Parser Generics: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseQuery}}; \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSponsor}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}};
  \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}
}

