% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parser-generics.R
\docType{methods}
\name{parseSponsor}
\alias{parseSponsor}
\title{LegiScan Parser Methods - parseSponsors}
\usage{
parseSponsor(theSponsor, dataframe = FALSE)
}
\arguments{
\item{theSponsor}{An object with the returned data from the sponsor method}

\item{dataframe}{If TRUE method returns a data.frame object;
if FALSE the method returns a tbl_df object}
}
\description{
Generic method for parsing data retrieved from getSponsors API call
}
\seealso{
Other LegiScan API Parser Generics: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseQuery}}; \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}}; \code{\link{parseSupplement}},
  \code{\link{parseSupplement,XMLDocumentContent-method}},
  \code{\link{parseSupplement,character-method}}
}

