% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parser-generics.R, R/parser-methods.R
\docType{methods}
\name{parseSupplement}
\alias{parseSupplement}
\alias{parseSupplement,XMLDocumentContent-method}
\alias{parseSupplement,character-method}
\title{LegiScan Parser Methods - parseSupplement}
\usage{
parseSupplement(theSupplement)

\S4method{parseSupplement}{XMLDocumentContent}(theSupplement)

\S4method{parseSupplement}{character}(theSupplement)
}
\arguments{
\item{theSupplement}{An object with the returned data from the supplement method}
}
\value{
A list object containing metadata and the MIME data retrieved by the API call
}
\description{
Generic method for parsing data retrieved from getSupplement API call

Method for parsing LegiScan API calls to getSupplement
}
\note{
The getSupplement API call method returns a binary MIME object with any
of several document format types.  Amendments and supplements are not supported
in all states.
}
\examples{
\dontrun{
# Create object of class LegiScan
myLegiScan <- legiscanR()

# Get the master list for Mississippi
theMasterList <- masterList(myLegiScan, "MS")

# Parse the XML/JSON formatted data into a data frame
cleanerMasterList <- parseMasterList(theMasterList)

# Get A bill (bill_id is in the 3rd Column of the master list)
aBill <- bill(myLegiScan, cleanerMasterList[1, 3])

# Parse the bill data
cleanerBill <- cleanBill(aBill, dataframe = FALSE, text = '')

# Retrieve the Supplement from the API call
aSupplement <- supplement(myLegiScan, cleanerBill[["supplements"]][["supplement_id"]])

# Parse the bill text response from LegiScan
cleanedSupplement <- parseSupplement(aSupplement)

}
}
\seealso{
Other LegiScan API Parser Generics: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseQuery}}; \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}}; \code{\link{parseSponsor}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}};
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}}

Other LegiScan Parser Methods: \code{\link{parseAmendment}},
  \code{\link{parseAmendment,XMLDocumentContent-method}},
  \code{\link{parseAmendment,character-method}};
  \code{\link{parseBillText}},
  \code{\link{parseBillText,XMLDocumentContent-method}},
  \code{\link{parseBillText,character-method}};
  \code{\link{parseBill}},
  \code{\link{parseBill,XMLDocumentContent,logical,character-method}},
  \code{\link{parseBill,character,logical,character-method}};
  \code{\link{parseMasterList}},
  \code{\link{parseMasterList,XMLDocumentContent,logical,logical-method}},
  \code{\link{parseMasterList,character,logical,logical-method}},
  \code{\link{parseSessions,character,logical,logical-method}};
  \code{\link{parseRollCall}},
  \code{\link{parseRollCall,XMLDocumentContent,logical-method}},
  \code{\link{parseRollCall,character,logical-method}};
  \code{\link{parseSessions}},
  \code{\link{parseSessions,XMLDocumentContent,logical-method}},
  \code{\link{parseSessions,character,logical-method}},
  \code{\link{parseStates}};
  \code{\link{parseSponsor,XMLDocumentContent,logical-method}},
  \code{\link{parseSponsor,character,logical-method}};
  \code{\link{parseStates,XMLDocumentContent,logical-method}},
  \code{\link{parseStates,character,logical-method}}
}

