% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constructor-generics.R, R/constructor-methods.R
\docType{methods}
\name{setUrl}
\alias{setUrl}
\alias{setUrl,LegiScan-method}
\title{LegiScan Constructor Methods - setUrl}
\usage{
setUrl(legiscan)

\S4method{setUrl}{LegiScan}(legiscan)
}
\arguments{
\item{legiscan}{An object of class LegiScan}
}
\value{
Returns the LegiScan class object with prespecified url string
}
\description{
Generic constructor method for setting the urlBase slot of a
LegiScan class object

Constructor method used for setting the urlBase slot
of a LegiScan class object
}
\note{
The URL string set by this method is "http://api.legiscan.com/?key="
}
\examples{
\dontrun{
# Create object of class LegiScan
myLegiScan <- legiscanR()

# Set the URL slot of the LegiScan object
setUrl(myLegiScan)
}
}
\seealso{
Other LegiScan Constructor Generics: \code{\link{setAPI}},
  \code{\link{setAPI,LegiScan,character,missing-method}},
  \code{\link{setAPI,LegiScan,missing,character-method}},
  \code{\link{setAPI,LegiScan,missing,missing-method}};
  \code{\link{setFormat}},
  \code{\link{setFormat,LegiScan,logical-method}}

Other LegiScan Constructor Methods: \code{\link{setAPI}},
  \code{\link{setAPI,LegiScan,character,missing-method}},
  \code{\link{setAPI,LegiScan,missing,character-method}},
  \code{\link{setAPI,LegiScan,missing,missing-method}};
  \code{\link{setFormat}},
  \code{\link{setFormat,LegiScan,logical-method}}
}

