% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caller-generics.R, R/caller-methods.R
\docType{methods}
\name{stateList}
\alias{stateList}
\alias{stateList,LegiScan-method}
\title{LegiScan Caller Methods - stateList}
\usage{
stateList(legiscan)

\S4method{stateList}{LegiScan}(legiscan)
}
\arguments{
\item{legiscan}{An object of class LegiScan}
}
\value{
Returns State List from LegiScan API method getStateList
}
\description{
Generic method for the LegiScan getStateList API call

A method to retrieve the list of States the user is authorized
to access from LegiScan
}
\examples{
\dontrun{
# Create new LegiScan object
myLegiScan <- legiscanR()

# Retrieve the State list available from the LegiScan API
stateList(myLegiScan)
}
}
\seealso{
Other LegiScan API Caller Generics: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,missing,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,missing,missing-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,numeric-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,missing,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{bill}},
  \code{\link{bill,LegiScan,missing,character,numeric-method}},
  \code{\link{bill,LegiScan,numeric,missing,missing-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,numeric-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,missing,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}
}

