% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xjformat.R
\name{xjformat}
\alias{xjformat}
\title{Determine if request was sent as XML or JSON}
\usage{
xjformat(x)
}
\arguments{
\item{x}{The result from the API call}
}
\value{
Returns the result object with additional class attributes
}
\description{
Checks the file heading to identify XML vs JSON content and appends the
appropriate class to the object.
}
\examples{
\dontrun{
# Use getURL from RCurl to make the API call
statelist <- getURL(requestURL)

# Determine if statelist object is XML or JSON
statelist <- xjformat(statelist)
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{billids}};
  \code{\link{fileLists}}; \code{\link{fileStructure}};
  \code{\link{getFullText}}; \code{\link{legiBillHist}};
  \code{\link{legiBillJSON}}; \code{\link{legiBillMeta}};
  \code{\link{legiBillProg}}; \code{\link{legiBillSpons}};
  \code{\link{legiBillTxt}}; \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}}; \code{\link{legiscanVotes}}
}

