% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caller-generics.R, R/caller-methods.R
\docType{methods}
\name{bill}
\alias{bill}
\alias{bill,LegiScan,missing,character,numeric-method}
\alias{bill,LegiScan,numeric,missing,missing-method}
\title{LegiScan Caller Methods - bill}
\usage{
bill(legiscan, id, state, billnumber)

\S4method{bill}{LegiScan,numeric,missing,missing}(legiscan, id, state,
  billnumber)

\S4method{bill}{LegiScan,missing,character,numeric}(legiscan, id, state,
  billnumber)
}
\arguments{
\item{legiscan}{An object of class LegiScan}

\item{id}{The LegiScan bill_id or LegiScan billnumber when called with a
valid state abbreviation}

\item{state}{A two character US State abbreviation (validated against the
state.abb dataset in R)}

\item{billnumber}{A LegiScan billnumber (can be retrieved in call to masterList)}
}
\description{
Generic method for the LegiScan getBill API call

Detailed data on specific bill including - but not limited to: sponsors,
full text (doc_id) identifier, and voting records identifiers

Detailed data on specific bill including - but not limited to: sponsors,
full text (doc_id) identifier, and voting records identifiers
}
\seealso{
Other LegiScan API Caller Generics: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,numeric-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,missing,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,numeric-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,missing,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}

Other LegiScan API Caller Methods: \code{\link{amendment}},
  \code{\link{amendment,LegiScan,numeric-method}};
  \code{\link{billText}},
  \code{\link{billText,LegiScan,numeric-method}};
  \code{\link{legisearch}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,missing,numeric-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,missing-method}},
  \code{\link{legisearch,LegiScan,character,missing,character,numeric,numeric-method}},
  \code{\link{legisearch,LegiScan,character,numeric,missing,missing,missing-method}};
  \code{\link{masterList}},
  \code{\link{masterList,LegiScan,character,missing-method}},
  \code{\link{masterList,LegiScan,missing,missing-method}};
  \code{\link{rollCall}},
  \code{\link{rollCall,LegiScan,numeric-method}};
  \code{\link{sessionList}},
  \code{\link{sessionList,LegiScan,character-method}};
  \code{\link{sponsor}},
  \code{\link{sponsor,LegiScan,numeric-method}};
  \code{\link{stateList}},
  \code{\link{stateList,LegiScan-method}};
  \code{\link{supplement}},
  \code{\link{supplement,LegiScan,numeric-method}}
}

