% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/billids.R
\name{billids}
\alias{billids}
\title{Bill Ids from LegiScan Bill API Requests}
\usage{
billids(billobject)
}
\arguments{
\item{billobject}{A file/API request to the bill data}
}
\value{
Returns a data.frame object of ID elements for LegiScan Bill data
}
\description{
Generates a data.frame object of ID elements for LegiScan Bill data
}
\examples{
\dontrun{
# Pass an unprocessed/cleaned bill object to billids to clean up the
# identifiers in the API response
billobject_cleaner <- billids(billobject_raw)
}
}
\seealso{
Other Parsing and Cleaning LegiScan Data: \code{\link{fileLists}};
  \code{\link{fileStructure}}; \code{\link{getFullText}};
  \code{\link{legiBillHist}}; \code{\link{legiBillJSON}};
  \code{\link{legiBillMeta}}; \code{\link{legiBillProg}};
  \code{\link{legiBillSpons}}; \code{\link{legiBillTxt}};
  \code{\link{legiPersonJSON}};
  \code{\link{legiVotesJSON}}; \code{\link{legiscanBill}};
  \code{\link{legiscanPerson}};
  \code{\link{legiscanVotes}}; \code{\link{xjformat}}
}

