% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset_list}
\alias{get_dataset_list}
\title{Get Dataset List}
\usage{
get_dataset_list(state = NULL, year = NULL)
}
\arguments{
\item{state}{Optional state abbreviation}

\item{year}{Optional year}
}
\description{
Get a list of all datasets
}
