% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legiscan-class.R
\docType{class}
\name{LegiScan-class}
\alias{LegiScan}
\alias{LegiScan-class}
\alias{initialize,LegiScan,character,character,character,...-method}
\alias{initialize,LegiScan,missing,character,character,...-method}
\alias{initialize,LegiScan,missing,character,missing,...-method}
\alias{initialize,LegiScan,missing,missing,character,...-method}
\alias{initialize,LegiScan,missing,missing,missing,...-method}
\alias{initialize,LegiScan-method}
\alias{legiscanR}
\title{LegiScan API Class object}
\usage{
legiscanR(urlBase = "http://api.legiscan.com/?key=",
  api = "ABCDEFGHIJKLMNOP", format = "JSON", ...)

\S4method{initialize}{LegiScan}(.Object, urlBase = "missing",
  api = "missing", format = "character", ...)

\S4method{initialize}{LegiScan}(.Object, urlBase = "missing",
  api = "missing", format = "character", ...)

\S4method{initialize}{LegiScan}(.Object, urlBase = "missing",
  api = "missing", format = "character", ...)

\S4method{initialize}{LegiScan}(.Object, urlBase = "missing",
  api = "missing", format = "character", ...)

\S4method{initialize}{LegiScan}(.Object, urlBase = "missing",
  api = "missing", format = "character", ...)
}
\arguments{
\item{urlBase}{Character string identifying root URL for API request}

\item{api}{LegiScan API key}

\item{format}{Character string indicating if requested data are in JSON or XML format}

\item{...}{Additional arguments (reserved for future use)}

\item{.Object}{Object of class LegiScan}
}
\description{
Defines a LegiScan class object which is passed to methods
to submit/retrieve data from calls to the LegiScan API and for parsing
data retrieve from the LegiScan API

LegiScan constructor function

LegiScan default construction method

LegiScan construction method for user specified format

LegiScan construction method for user specified format

LegiScan construction method with user specified api key value/path

LegiScan construction method with user specified api and format values
}
\section{Slots}{

\describe{
\item{\code{urlBase}}{The base for the request URL (predefined in the class prototype)}

\item{\code{api}}{The API key for the end user}

\item{\code{format}}{Specifies if you have requested your API pulls in XML or JSON}
}}
\note{
The default behavior for the class constructor is to read the API key
from ~/.legiscanAPI.  Accessor methods are available to get/set the values
of the LegiScan object.

The file format returned by the LegiScan API is set up when logging
into your LegiScan account.  The value provided by the class object is used
to select the appropriate parsing methods based on the object returned
from the API call

The api value should either be the file path to the .legiscanAPI file
or be the api key itself

The api value should either be the file path to the .legiscanAPI file
or be the api key itself
}
\examples{
\donttest{
# Create new object with class LegiScan
x <- legiscanR()

# Check class of the object x
class(x)
}
}

